requirements.txt
flask
pydantic
pandas
numpy
scikit-learn
shap
PyJWT
